requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306; platform_system == "Windows"
six==1.16.0
SQLAlchemy==2.0.28
sqlalchemy-access==2.0.2; platform_system == "Windows"
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
